// src/lib/api/generated.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/api/v1/shelters/results": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get: operations["getApiV1SheltersResults"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/api/v1/shelters/contribute": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    post: operations["postApiV1SheltersContribute"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/api/v1/shelters/": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get: operations["getApiV1Shelters"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: never;
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
  getApiV1SheltersResults: {
    parameters: {
      query?: {
        search?: string | null;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description Success response */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": {
            /** @example 4 */
            contributors: number;
            /**
             * Format: float
             * @example 7.5
             */
            contribution: number | null;
          };
        };
      };
    };
  };
  postApiV1SheltersContribute: {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: {
      content: {
        "application/json": {
          contributors: {
            /** @example Jan */
            firstName: string;
            /** @example Dobry */
            lastName: string;
            /**
             * Format: email
             * @example email@email.com
             */
            email: string;
            /** @example 0900 000 000 */
            phone?: string | null;
          }[];
          /**
           * Format: float
           * @example 1
           */
          shelterID?: number | null;
          /**
           * Format: float
           * @example 1.5
           */
          value: number;
        };
      };
    };
    responses: {
      /** @description Success response */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": {
            messages: {
              /** @example Príspevok bol úspešne zaznamenaný */
              message: string;
              /**
               * @example SUCCESS
               * @enum {string}
               */
              type: "ERROR" | "WARNING" | "INFO" | "SUCCESS";
            }[];
          };
        };
      };
    };
  };
  getApiV1Shelters: {
    parameters: {
      query?: {
        search?: string | null;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description Success response */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": {
            shelters?: {
              /** @example 4 */
              id: number;
              /** @example Útulok pre psov - TEZAS */
              name: string;
            }[];
          };
        };
      };
    };
  };
}